+++ app/(root)/page.tsx
@@ -1,15 +1,15 @@
-import CardBenefits from '@components/CardBenefits';
-import CardPlans from '@components/CardPlans';
+import Benefits from '@components/Benefits';
 import Form from '@components/commons/Form';
 import Introduce from '@components/Introduce';
+import Plans from '@components/Plans';
 
 export default function Home() {
   return (
     <main>
       <Introduce />
       <Form />
-      <CardBenefits />
-      <CardPlans />
+      <Benefits />
+      <Plans />
     </main>
   );
 }
