+++ app/components/Plans/motion.ts
@@ -24,3 +24,13 @@ export const inVariants = {
     },
   },
 };
+
+export const iconVariants = {
+  hidden: {
+    opacity: 0,
+  },
+  show: {
+    opacity: 1,
+    rotate: [0, 0, 45, -25, 0],
+  },
+};
